refactor(ui): extract InfoTooltip component and remove legacy InfoField
