+++ ui/app/(prowler)/attack-paths/(workflow)/query-builder/_components/node-detail/node-overview.tsx
@@ -1,7 +1,7 @@
 "use client";
 
+import { InfoField } from "@/components/shadcn/info-field/info-field";
 import { CodeSnippet } from "@/components/ui/code-snippet/code-snippet";
-import { InfoField } from "@/components/ui/entities";
 import { DateWithTime } from "@/components/ui/entities/date-with-time";
 import type { GraphNode, GraphNodePropertyValue } from "@/types/attack-paths";
